refactor(workstation-pr): split useWorkstationPrDetail into modules
Move the React-free fetch layer (including the single module-level
in-flight dedup map) into workstationPrDetailFetch, the write paths into
useWorkstationPrMutations, reviewer candidates into
useWorkstationPrReviewers, and upsertById/bumpRequestId into the
existing workstationPrHelpers. Behaviour unchanged.
